+++ tokio-signal/CHANGELOG.md
@@ -1,6 +1,9 @@
-# 0.2.9
+# 0.2.9 (February 4, 2020)
+
+* Add `tokio 0.2.x` deprecation notice.
 
 ### Fixed
+
 - `windows::Event` performs internal registrations lazily, so now it can be
 constructed outside of a running task
 - remove usage of deprecated `Handle::current` in default `windows::Event`